requirements.txt
tensorflow<2.11
numpy==1.26.4
pandas
pillow
scikit-learn==1.2.2